+++ src/components/FormContainer/Error.tsx
@@ -1,6 +1,6 @@
 import { HTMLAttributes } from 'react';
 import { cn } from '@/lib/cva';
-import styles from './FormContainer.module.css';
+import styles from './Error.module.css';
 
 export const Error = ({ className, ...props }: HTMLAttributes<HTMLDivElement>) => (
   <div
