--- src/components/FormContainer/FormContainer.module.css
@@ -1,53 +1,0 @@
-.form-root {
-  display: flex;
-  width: 100%;
-  flex-direction: column-reverse;
-  align-items: flex-start;
-  gap: var(--click-field-space-gap);
-}
-
-.form-root * {
-  outline: none;
-  box-shadow: none;
-}
-
-.form-root_direction_column-reverse {
-  flex-direction: column-reverse;
-}
-
-.form-root_direction_column {
-  flex-direction: column;
-}
-
-.form-root_direction_row-reverse {
-  flex-direction: row-reverse;
-}
-
-.form-root_direction_row {
-  flex-direction: row;
-}
-
-.form-root_add-label-padding label {
-  padding-top: calc(var(--click-field-space-y) + 1px);
-  line-height: 1lh;
-}
-
-.form-element-container {
-  display: flex;
-  /* stylelint-disable plugin/no-unsupported-browser-features, declaration-property-value-no-unknown -- the original rule used this same
-   width fallback chain (100% → -webkit-fill-available → fill-available →
-   stretch) so the container fills the available inline space. */
-  width: 100%;
-  width: -webkit-fill-available;
-  width: fill-available;
-  width: stretch;
-  /* stylelint-enable plugin/no-unsupported-browser-features, declaration-property-value-no-unknown */
-  flex-direction: column;
-  align-items: flex-start;
-  gap: var(--click-field-space-gap);
-}
-
-.error {
-  color: var(--click-field-color-label-error);
-  font: var(--click-field-typography-label-error);
-}
