+++ src/components/FormContainer/FormElementContainer.tsx
@@ -1,6 +1,6 @@
 import { HTMLAttributes } from 'react';
 import { cn } from '@/lib/cva';
-import styles from './FormContainer.module.css';
+import styles from './FormElementContainer.module.css';
 
 export const FormElementContainer = ({
   className,
